+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [Unreleased]
+
+### Features
+
+- Add user flagging support for moderation
+  - Add `flagUser()` method to `StreamClient` class
+  - Add `flag()` method to `StreamUser` class
+  - Add `FlagUserOptions` and `FlagAPIResponse` TypeScript types
+
 ## [8.8.0](https://github.com/GetStream/stream-js/compare/v8.7.0...v8.8.0) (2025-04-10)
 
 ## [8.7.0](https://github.com/GetStream/stream-js/compare/v8.6.1...v8.7.0) (2025-03-14)